refactor: delete deprecated attributess, move connect() in parent class
